Decompose analysis function for datalog reading.
